+++ R/tree_information.R
@@ -80,13 +80,6 @@ CladisticInfo <- function(x) UseMethod("CladisticInfo")
 
 #' @rdname CladisticInfo
 #' @export
-PhylogeneticInfo <- function(x) {                                               # nocov start
-  .Deprecated("CladisticInfo()")
-  UseMethod("CladisticInfo")
-}                                                                               # nocov end
-
-#' @rdname CladisticInfo
-#' @export
 CladisticInfo.phylo <- function(x) {
   Log2Unrooted(NTip(x)) - Log2TreesMatchingTree(x)
 }
@@ -105,10 +98,6 @@ CladisticInfo.list <- function(x) vapply(x, CladisticInfo, 0)
 #' @export
 CladisticInfo.multiPhylo <- CladisticInfo.list
 
-
-#' @rdname CladisticInfo
-#' @export
-PhylogeneticInformation <- PhylogeneticInfo
 #' @rdname CladisticInfo
 #' @export
 CladisticInformation <- CladisticInfo