+++ NEWS.md
@@ -4,7 +4,7 @@
 - Fix `RoguePlot(sort = TRUE)`
   [Rogue#33](https://github.com/ms609/Rogue/issues/33).
 - Remove R.cache dependency: `UnrootedKeys()` now uses a native cache implementation.
-- Remove deprecated function `TreeSplits()`.
+- Remove deprecated functions `TreeSplits()`, `ForestSplits()`, `SplitNumber()`.
 
 
 # TreeTools 1.16.1 (2025-08-24) #