+++ NEWS.md
@@ -1,3 +1,6 @@
+# TreeTools 2.0.0.9000 (development) #
+- (miscellaneous tidying)
+
 # TreeTools 2.0.0 (2025-09-23) #
 
 ## New functionality